Opdateret dosis til float i hele programmet. Mangler styrke til float i print_recept function

+++ src/main.h
@@ -33,5 +33,5 @@ int medicine_callback(void *, int, char **, char **);
 
 int check_med(char med_input[20]);
 int check_med_max(char med_input[20]);
-void insert_recept(int medid, int dosis, int frek, char desk[250]);
+void insert_recept(int medid, double dosis, int frek, char desk[250]);
 void delete_recept(int valg);
